+++ android/src/main/java/com/applovin/reactnative/AppLovinMAXModule.java
@@ -151,6 +151,7 @@ public class AppLovinMAXModule
     // Store these values if pub attempts to set it before initializing
     private       List<String>        initializationAdUnitIdsToSet;
     private       String              userIdToSet;
+    private       Boolean             mutedToSet;
     private       List<String>        testDeviceAdvertisingIdsToSet;
     private       Boolean             verboseLoggingToSet;
     private       Boolean             creativeDebuggerEnabledToSet;
@@ -329,6 +330,13 @@ public class AppLovinMAXModule
         {
             settings.getTermsAndPrivacyPolicyFlowSettings().setDebugUserGeography( getAppLovinConsentFlowUserGeography( debugUserGeographyToSet ) );
             debugUserGeographyToSet = null;
+        }
+
+        // Set muted if needed
+        if ( mutedToSet != null )
+        {
+            settings.setMuted( mutedToSet );
+            mutedToSet = null;
         }
 
         // Set test device ids if needed
@@ -531,9 +539,15 @@ public class AppLovinMAXModule
     @ReactMethod
     public void setMuted(final boolean muted)
     {
-        if ( !isPluginInitialized ) return;
-
-        sdk.getSettings().setMuted( muted );
+        if ( isPluginInitialized )
+        {
+            sdk.getSettings().setMuted( muted );
+            mutedToSet = null;
+        }
+        else
+        {
+            mutedToSet = muted;
+        }
     }
 
     @ReactMethod